This is a C program for automating the billing system in a medical store.
The shopkeeper will have to enter the details of the items that were bought into the shop newly and these details are stored in a file which is double indexed with two index files primary index file and a secondary index file.
When a customer asks for a medicine the shopkeeper can search for it in the system by its name or its major component and sell the item by generating a bill. This sale is updated in the file to reflect the change in the number of items left.
the shopkeeper has to select the dispose option once in every month so that the expired items are removed out from  the system.